+++ css/style.css
@@ -311,6 +311,7 @@ header button:hover {
     padding: 0.5rem;
     
 }
+ 
 
 .desempenho {
     width: 100%;
@@ -394,12 +395,38 @@ header button:hover {
     width: 100%;
     margin-top: 4rem;
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
+    justify-content: space-between;
 }
 
-.ranking-temas span {
-    text-transform: lowercase;
-    font-weight: 200;
+.conteiner-ranking{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border: 5px solid var(--font-color-3);
+    padding: 2rem;
+    width: 200px;
+    border-radius: 15px;
+}
+
+.ranking-temas ol{
+    display: flex;
+    flex-direction: column;
+    margin-top: 20px;
+    border: none;
+    gap: 1rem;
+    color: var(--font-color-3);
+}
+
+.ranking-temas ol li div{
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    gap: 1.5rem;
+}
+
+.ranking-temas p, h4 {
+    font-weight: 600;
     font-size: 1rem;
     color: var(--font-color-3);
 }
