+++ index.html
@@ -62,15 +62,15 @@
         </div>
 
         <div class="ranking-temas">
-            <div>
+            <div class="conteiner-ranking">
                 <h3>HTML</h3>
                 <ol class="ranking-temas-html"></ol>
             </div>
-            <div>
+            <div class="conteiner-ranking">
                 <h3>CSS</h3>
                 <ol class="ranking-temas-css"></ol>
             </div>
-            <div>
+            <div class="conteiner-ranking">
                 <h3>JavaScript</h3>
                 <ol class="ranking-temas-javascript"></ol>
             </div>
@@ -86,13 +86,7 @@
         <button id="btn-concluir" class="btn">Concluir</button>
         <button id="btn-jogar" class="btn">Jogar novamente</button>
     </div> 
-    
-    <script type="module" src="./js/cronometro.js"></script>
-    <script type="module" src="./js/musica.js"></script>
-    <script type="module" src="./js/mudancaTema.js"></script>
-    <script type="module" src="./js/exibePerguntas.js"></script>
-    <script type="module" src="./js/tela-inicial.js"></script>
-    <script type="module" src="./js/verificarRespostas.js"></script>
-    <script type="module" src="./js/resultadoQuiz.js"></script>
+
+    <script type="module" src="./js/script.js"></script>
 </body>
 </html>
