+++ js/dadosRanking.js
@@ -14,5 +14,37 @@ export const dadosRanking = [
         tempo: "00:30",
         data: "29/07/2023",
         pontos: 7
+    },
+    {
+        id: 3,
+        nome: "Isabela",
+        tema: "HTML",
+        tempo: "00:55",
+        data: "30/07/2023",
+        pontos: 8
+    },
+    {
+        id: 4,
+        nome: "Rafaela",
+        tema: "JavaScript",
+        tempo: "01:10",
+        data: "31/07/2023",
+        pontos: 10
+    },
+    {
+        id: 5,
+        nome: "Marcos",
+        tema: "HTML",
+        tempo: "00:50",
+        data: "01/08/2023",
+        pontos: 10
+    },
+    {
+        id: 6,
+        nome: "João",
+        tema: "CSS",
+        tempo: "00:25",
+        data: "02/08/2023",
+        pontos: 5
     }
 ];
