+++ js/mudancaTema.js
@@ -1,17 +1,19 @@
 //alterar o tema para claro ou escuro
 
-const tema = document.getElementById('tema')
-const botao = document.getElementById('toggle-theme')
-const corpo = document.getElementById('corpo')
-
-botao.addEventListener('click', () => {
-    if(tema.innerText=="light_mode"){
-        tema.innerText="dark_mode"
-        corpo.setAttribute('class', 'light')
-    }
-    else{
-        tema.innerText="light_mode" 
-        corpo.setAttribute('class', 'dark')
-    }
-
-})
+export function darkMode(){
+    const tema = document.getElementById('tema')
+    const botao = document.getElementById('toggle-theme')
+    const corpo = document.getElementById('corpo')
+    
+    botao.addEventListener('click', () => {
+        if(tema.innerText=="light_mode"){
+            tema.innerText="dark_mode"
+            corpo.setAttribute('class', 'light')
+        }
+        else{
+            tema.innerText="light_mode" 
+            corpo.setAttribute('class', 'dark')
+        }
+    
+    })
+}
